+++ src/EventListener/ParseArticlesHook.php
@@ -1,16 +1,14 @@
 <?php
-/**
- * Contao Open Source CMS
+
+/*
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
  *
- * Copyright (c) 2017 Heimrich & Hannot GmbH
- *
- * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
- * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ * @license LGPL-3.0-or-later
  */
 
 namespace HeimrichHannot\ContaoDisqusBundle\EventListener;
 
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\Environment;
 use Contao\FrontendTemplate;
 use Contao\Module;
@@ -21,7 +19,7 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 class ParseArticlesHook
 {
     /**
-     * @var ContaoFrameworkInterface
+     * @var ContaoFramework
      */
     private $framework;
     /**
@@ -33,38 +31,33 @@ class ParseArticlesHook
      */
     private $urlGenerator;
 
-    public function __construct(ContaoFrameworkInterface $framework, DisqusRenderer $renderer, UrlGeneratorInterface $urlGenerator)
+    public function __construct(ContaoFramework $framework, DisqusRenderer $renderer, UrlGeneratorInterface $urlGenerator)
     {
-        $this->framework    = $framework;
-        $this->renderer     = $renderer;
+        $this->framework = $framework;
+        $this->renderer = $renderer;
         $this->urlGenerator = $urlGenerator;
     }
-
 
     public function addDisqus(FrontendTemplate &$template, array $article, Module $module)
     {
         global $objPage;
         /** @var NewsArchiveModel $newsArchive */
         $newsArchive = $this->framework->getAdapter(NewsArchiveModel::class)->findById($article['pid']);
-        if (!$newsArchive->disqus_addDisqus)
-        {
+        if (!$newsArchive->disqus_addDisqus) {
             return;
         }
         $disqus_shortname = $newsArchive->disqus_shortname;
-        if ($newsArchive->disqus_identifier)
-        {
+        if ($newsArchive->disqus_identifier) {
             $disqus_identifier = str_replace('{id}', $article['id'], $newsArchive->disqus_identifier);
-        } else
-        {
+        } else {
             $disqus_identifier = $article['id'];
         }
 
         $result = $this->renderer->render($disqus_shortname, $disqus_identifier);
 
-        $url                      = $this->framework->getAdapter(Environment::class)->get('url');
-        $path                     = $this->urlGenerator->generate($objPage->id);
-        $disqus_pageUrl           = $url . '/' . $path . '/' . $article['alias'];
+        $url = $this->framework->getAdapter(Environment::class)->get('url');
+        $path = $this->urlGenerator->generate($objPage->id);
+        $disqus_pageUrl = $url.'/'.$path.'/'.$article['alias'];
         $template->disqus_section = $this->renderer->render($disqus_shortname, $disqus_identifier, $disqus_pageUrl);
-        return;
     }
 }