+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased] - 2022-10-11
+- Changed: allow php 8
+- Fixed: compatibility with contao 4.13
 - Fixed: added missing license file
 
 ## [2.2.1] - 2022-01-12